# Reduce memory churn, avoid generating ctx when not necessary, much faster valid primitives