--- a/deno/lib/helpers/parseUtil.ts
+++ b/deno/lib/helpers/parseUtil.ts
@@ -125,11 +125,22 @@ export type ParsePath = ParsePathComponent[];
 export const EMPTY_PATH: ParsePath = [];
 
 export interface ParseContext {
+<<<<<<< HEAD
   readonly path: ParsePath;
   readonly issues: ZodIssue[];
   readonly schemaErrorMap?: ZodErrorMap;
   readonly contextualErrorMap?: ZodErrorMap;
   readonly async: boolean;
+=======
+  readonly common: {
+    readonly issues: ZodIssue[];
+    readonly contextualErrorMap?: ZodErrorMap;
+    readonly async: boolean;
+    readonly typeCache: Map<any, ZodParsedType> | undefined;
+  };
+  readonly path: ParsePath;
+  readonly schemaErrorMap?: ZodErrorMap;
+>>>>>>> 4404b27 (Consolidate common elements of ctx. Remove spreads.)
   readonly parent: ParseContext | null;
   readonly data: any;
   readonly parsedType: ZodParsedType;
@@ -150,13 +161,21 @@ export function addIssueToContext(
     data: ctx.data,
     path: ctx.path,
     errorMaps: [
+<<<<<<< HEAD
       ctx.contextualErrorMap, // contextual error map is first priority
+=======
+      ctx.common.contextualErrorMap, // contextual error map is first priority
+>>>>>>> 4404b27 (Consolidate common elements of ctx. Remove spreads.)
       ctx.schemaErrorMap, // then schema-bound map if available
       overrideErrorMap, // then global override map
       defaultErrorMap, // then global default map
     ].filter((x) => !!x) as ZodErrorMap[],
   });
+<<<<<<< HEAD
   ctx.issues.push(issue);
+=======
+  ctx.common.issues.push(issue);
+>>>>>>> 4404b27 (Consolidate common elements of ctx. Remove spreads.)
 }
 
 export type ObjectPair = {