--- a/src/helpers/parseUtil.ts
+++ b/src/helpers/parseUtil.ts
@@ -125,11 +125,14 @@ export type ParsePath = ParsePathComponent[];
 export const EMPTY_PATH: ParsePath = [];
 
 export interface ParseContext {
+  readonly common: {
+    readonly issues: ZodIssue[];
+    readonly contextualErrorMap?: ZodErrorMap;
+    readonly async: boolean;
+    readonly typeCache: Map<any, ZodParsedType> | undefined;
+  };
   readonly path: ParsePath;
-  readonly issues: ZodIssue[];
   readonly schemaErrorMap?: ZodErrorMap;
-  readonly contextualErrorMap?: ZodErrorMap;
-  readonly async: boolean;
   readonly parent: ParseContext | null;
   readonly data: any;
   readonly parsedType: ZodParsedType;
@@ -150,13 +153,13 @@ export function addIssueToContext(
     data: ctx.data,
     path: ctx.path,
     errorMaps: [
-      ctx.contextualErrorMap, // contextual error map is first priority
+      ctx.common.contextualErrorMap, // contextual error map is first priority
       ctx.schemaErrorMap, // then schema-bound map if available
       overrideErrorMap, // then global override map
       defaultErrorMap, // then global default map
     ].filter((x) => !!x) as ZodErrorMap[],
   });
-  ctx.issues.push(issue);
+  ctx.common.issues.push(issue);
 }
 
 export type ObjectPair = {