… coverage.svg
… src/types.ts