Okay, last one of these, hope they aren't annoying 😆 

So, `._processInputParams` is called pretty uniformly throughout `types.ts`, but in many cases the `ctx` and `status` values it produces are never used, so they only contribute to overhead and memory churn. This PR defers calling processInputParams: in cases like numbers in which ctx is useful, it's generated lazily, only when necessary.

Ctx is totally still useful for collection types - unions, objects, arrays - so it is still used there. But at the leaf values like strings & numbers, deferring or avoiding creating context and parserstate objects adds up to very respectable gains 💪

- Strings: 2,572,767 ops/sec to 8,020,277 ops/sec
- Literals: 2,764,958 ops/sec to 13,257,671 ops/sec
- Enums: 3,023,900 to 25,570,525 ops/sec
- Numbers: 2,151,291 to 6,242,437

These are just for the "valid" cases. Invalid results have stayed the same - we still generate a ctx when it's necessary.

<details>
<summary>Before</summary>
<pre>
yarn run v1.22.4
$ ts-node src/benchmarks/index.ts
z.enum: valid x 3,023,900 ops/sec ±0.50% (95 runs sampled)
z.enum: invalid x 88,480 ops/sec ±1.33% (95 runs sampled)
z.undefined: valid x 2,763,887 ops/sec ±0.67% (94 runs sampled)
z.undefined: invalid x 92,613 ops/sec ±0.78% (95 runs sampled)
z.literal: valid x 2,764,958 ops/sec ±0.65% (97 runs sampled)
z.literal: invalid x 92,877 ops/sec ±0.62% (94 runs sampled)
z.number: valid x 2,151,291 ops/sec ±19.28% (92 runs sampled)
z.number: invalid type x 93,060 ops/sec ±0.71% (91 runs sampled)
z.number: invalid number x 90,584 ops/sec ±0.62% (94 runs sampled)
z.string: empty string x 2,298,812 ops/sec ±5.68% (91 runs sampled)
z.string: short string x 2,350,511 ops/sec ±0.43% (94 runs sampled)
z.string: long string x 2,338,882 ops/sec ±0.43% (95 runs sampled)
z.string: invalid: null x 90,197 ops/sec ±4.11% (90 runs sampled)
z.string: manual parser: long x 879,926,450 ops/sec ±1.59% (86 runs sampled)
z.object: empty: valid x 1,689,127 ops/sec ±1.00% (94 runs sampled)
z.object: empty: valid: extra keys x 1,603,788 ops/sec ±0.51% (94 runs sampled)
z.object: empty: invalid: null x 90,512 ops/sec ±4.32% (91 runs sampled)
z.object: short: valid x 829,616 ops/sec ±0.95% (95 runs sampled)
z.object: short: valid: extra keys x 796,218 ops/sec ±0.48% (95 runs sampled)
z.object: short: invalid: null x 90,538 ops/sec ±3.58% (94 runs sampled)
z.object: long: valid x 416,655 ops/sec ±0.42% (95 runs sampled)
z.object: long: valid: extra keys x 409,889 ops/sec ±0.33% (94 runs sampled)
z.object: long: invalid: null x 92,829 ops/sec ±0.63% (94 runs sampled)
z.union: double: valid: a x 258,928 ops/sec ±6.25% (89 runs sampled)
z.union: double: valid: b x 266,564 ops/sec ±0.88% (93 runs sampled)
z.union: double: invalid: null x 34,444 ops/sec ±0.55% (92 runs sampled)
z.union: double: invalid: wrong shape x 31,241 ops/sec ±0.72% (96 runs sampled)
z.union: many: valid: a x 125,748 ops/sec ±6.13% (89 runs sampled)
z.union: many: valid: c x 130,767 ops/sec ±0.87% (96 runs sampled)
z.union: many: invalid: null x 21,682 ops/sec ±0.54% (95 runs sampled)
z.union: many: invalid: wrong shape x 19,372 ops/sec ±0.72% (93 runs sampled)
z.discriminatedUnion: double: valid: a x 628,211 ops/sec ±4.38% (93 runs sampled)
z.discriminatedUnion: double: valid: b x 649,602 ops/sec ±0.62% (93 runs sampled)
z.discriminatedUnion: double: invalid: null x 88,935 ops/sec ±0.44% (97 runs sampled)
z.discriminatedUnion: double: invalid: wrong shape x 79,578 ops/sec ±4.03% (90 runs sampled)
z.discriminatedUnion: many: valid: a x 647,738 ops/sec ±0.57% (95 runs sampled)
z.discriminatedUnion: many: valid: c x 635,873 ops/sec ±0.60% (94 runs sampled)
z.discriminatedUnion: many: invalid: null x 89,762 ops/sec ±0.40% (95 runs sampled)
z.discriminatedUnion: many: invalid: wrong shape x 77,317 ops/sec ±3.93% (91 runs sampled)
✨  Done in 213.99s.
</pre>
</details>

<details>
<summary>After</summary>
<pre>
yarn run v1.22.4
$ ts-node src/benchmarks/index.ts
z.enum: valid x 25,570,525 ops/sec ±0.58% (96 runs sampled)
z.enum: invalid x 88,718 ops/sec ±0.79% (95 runs sampled)
z.undefined: valid x 9,202,311 ops/sec ±0.78% (95 runs sampled)
z.undefined: invalid x 91,320 ops/sec ±0.77% (93 runs sampled)
z.literal: valid x 13,257,671 ops/sec ±0.97% (92 runs sampled)
z.literal: invalid x 91,489 ops/sec ±4.37% (91 runs sampled)
z.number: valid x 6,242,437 ops/sec ±0.44% (93 runs sampled)
z.number: invalid type x 91,081 ops/sec ±0.82% (94 runs sampled)
z.number: invalid number x 89,370 ops/sec ±0.94% (94 runs sampled)
z.string: empty string x 7,118,336 ops/sec ±5.52% (89 runs sampled)
z.string: short string x 7,601,530 ops/sec ±1.06% (94 runs sampled)
z.string: long string x 7,644,857 ops/sec ±0.45% (93 runs sampled)
z.string: invalid: null x 91,542 ops/sec ±3.58% (90 runs sampled)
z.string: manual parser: long x 885,283,907 ops/sec ±1.77% (87 runs sampled)
z.object: empty: valid x 1,650,805 ops/sec ±0.63% (95 runs sampled)
z.object: empty: valid: extra keys x 1,525,619 ops/sec ±0.62% (94 runs sampled)
z.object: empty: invalid: null x 90,888 ops/sec ±4.49% (90 runs sampled)
z.object: short: valid x 1,107,392 ops/sec ±0.64% (92 runs sampled)
z.object: short: valid: extra keys x 1,094,233 ops/sec ±0.49% (95 runs sampled)
z.object: short: invalid: null x 96,996 ops/sec ±1.03% (94 runs sampled)
z.object: long: valid x 507,361 ops/sec ±20.42% (94 runs sampled)
z.object: long: valid: extra keys x 554,007 ops/sec ±0.54% (93 runs sampled)
z.object: long: invalid: null x 96,142 ops/sec ±0.89% (94 runs sampled)
z.union: double: valid: a x 280,189 ops/sec ±5.72% (93 runs sampled)
z.union: double: valid: b x 284,643 ops/sec ±0.84% (92 runs sampled)
z.union: double: invalid: null x 33,330 ops/sec ±1.10% (93 runs sampled)
z.union: double: invalid: wrong shape x 30,885 ops/sec ±0.73% (93 runs sampled)
z.union: many: valid: a x 132,640 ops/sec ±5.59% (94 runs sampled)
z.union: many: valid: c x 136,428 ops/sec ±0.39% (94 runs sampled)
z.union: many: invalid: null x 20,957 ops/sec ±0.61% (92 runs sampled)
z.union: many: invalid: wrong shape x 18,856 ops/sec ±1.07% (91 runs sampled)
z.discriminatedUnion: double: valid: a x 793,625 ops/sec ±5.48% (91 runs sampled)
z.discriminatedUnion: double: valid: b x 819,445 ops/sec ±0.57% (94 runs sampled)
z.discriminatedUnion: double: invalid: null x 90,611 ops/sec ±0.76% (94 runs sampled)
z.discriminatedUnion: double: invalid: wrong shape x 82,642 ops/sec ±3.94% (90 runs sampled)
z.discriminatedUnion: many: valid: a x 804,635 ops/sec ±0.83% (94 runs sampled)
z.discriminatedUnion: many: valid: c x 810,716 ops/sec ±0.80% (94 runs sampled)
z.discriminatedUnion: many: invalid: null x 89,417 ops/sec ±0.84% (94 runs sampled)
z.discriminatedUnion: many: invalid: wrong shape x 79,459 ops/sec ±3.90% (93 runs sampled)
✨  Done in 214.84s
</pre>
</details>