@colinhacks 2022-03-18
Beautiful, looks great. While poking around this PR I saw a couple other easy wins that should help with the memory issue so tacked those on here.